+++ src/game/collision/CustomBlockBox.java
@@ -2,41 +2,40 @@ package game.collision;
 
 public class CustomBlockBox {
 
-    private static float[] aabb = new float[6];
+    private static double[] aabb = new double[6];
 
-    public static void setBlockBox(int x, int y, int z, float[] blockBox){
+    public static void setBlockBox(int x, int y, int z, double[] blockBox){
         aabb[0] = x + blockBox[0];
         aabb[1] = y + blockBox[1];
         aabb[2] = z + blockBox[2];
         aabb[3] = x + blockBox[3];
         aabb[4] = y + blockBox[4];
         aabb[5] = z + blockBox[5];
-
     }
 
     //getters
 
-    public static float BlockBoxGetLeft(){
+    public static double BlockBoxGetLeft(){
         return aabb[0];
     }
 
-    public static float BlockBoxGetBottom(){
+    public static double BlockBoxGetBottom(){
         return aabb[1];
     }
 
-    public static float BlockBoxGetFront(){
+    public static double BlockBoxGetFront(){
         return aabb[2];
     }
 
-    public static float BlockBoxGetRight(){
+    public static double BlockBoxGetRight(){
         return aabb[3];
     }
 
-    public static float BlockBoxGetTop(){
+    public static double BlockBoxGetTop(){
         return aabb[4];
     }
 
-    public static float BlockBoxGetBack(){
+    public static double BlockBoxGetBack(){
         return aabb[5];
     }
 }
